requirements.txt
python-dotenv 
langchain 
langchain-openai 
TTS
requests 
streamlit 
